pyproject.toml
[project]
name = "test-task-django-rishat"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitry",email = "kiselevdv1991@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "stripe (>=11.6.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-bootstrap5 (>=24.3,<25.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
isort = "^6.0.0"

